+++ pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
   cupertino_icons: ^0.1.2
   flutter_staggered_animations: ^0.1.2
   provider: ">=4.0.0 <5.0.0"
+  responsive_builder: ^0.1.2
 
 dev_dependencies:
   flutter_test:
